+++ tests/test_checker/test_noqa.py
@@ -123,6 +123,7 @@ def test_noqa_fixture_disabled(absolute_path, all_violations):
         'Z451': 2,
         'Z452': 2,
         'Z453': 1,
+        'Z454': 1,
     }
 
     process = subprocess.Popen(